feat: Move freelancer-client flow to jobDetail page
